Changes to file structure, and added a new login screen
