--- app/src/main/java/com/mylesspencertyler/snuber/MainActivity.java
@@ -1,45 +1,0 @@
-package com.mylesspencertyler.snuber;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
-
-import com.mylesspencertyler.snuber.utils.Utils;
-
-public class MainActivity extends AppCompatActivity {
-
-    private static final String TAG = MainActivity.class.getSimpleName();
-
-    Toolbar toolbar;
-    Intent intent;
-
-    String[] colors = {"#96CC7A", "#EA705D", "#66BBCC"};
-
-    public static final String PREF_USER_FIRST_TIME = "user_first_time";
-    boolean isUserFirstTime;
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-
-
-    }
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        isUserFirstTime = Boolean.valueOf(Utils.readSharedSetting(MainActivity.this, PREF_USER_FIRST_TIME, "true"));
-
-        Intent introIntent = new Intent(MainActivity.this, PagerActivity.class);
-        introIntent.putExtra(PREF_USER_FIRST_TIME, isUserFirstTime);
-
-        if (isUserFirstTime)
-            startActivity(introIntent);
-
-
-        setContentView(R.layout.activity_main);
-    }
-}
