+++ app/src/main/java/com/mylesspencertyler/snuber/activity/SplashActivity.java
@@ -1,4 +1,4 @@
-package com.mylesspencertyler.snuber;
+package com.mylesspencertyler.snuber.activity;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -10,7 +10,7 @@ public class SplashActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        Intent intent = new Intent(this, MainActivity.class);
+        Intent intent = new Intent(this, LoginActivity.class);
         startActivity(intent);
         finish();
     }
