+++ app/src/main/res/layout/activity_pager.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.mylesspencertyler.snuber.PagerActivity">
+    tools:context="com.mylesspencertyler.snuber.activity.PagerActivity">
 
         <android.support.v4.view.ViewPager
             android:id="@+id/container"
