+++ app/src/main/res/layout/fragment_pager.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="@dimen/activity_horizontal_margin"
-    tools:context="com.mylesspencertyler.snuber.PagerActivity$PlaceholderFragment">
+    tools:context="com.mylesspencertyler.snuber.activity.PagerActivity$PlaceholderFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
